requirements-vendor.txt
Django==1.9
djangorestframework
